+++ CHANGELOG.md
@@ -1,1 +1,7 @@
 # TransMate Changelog
+
+## Unreleased
+### Changed
+- TransMate now uses the `Field::EVENT_DEFINE_ACTION_MENU_ITEMS` event to add the "Translate field" action
+### Fixed
+- Fixed a bug where TransMate would prevent all field layout element conditions from working 🙈